Typescript compiler config added install.sh file created and working Typescript bifercation done
