+++ backend-service/lib-js/paths.js
@@ -1,0 +1,6 @@
+import {join , dirname} from 'path';
+
+
+const LOGS_DIRECTORY = join(dirname("."),"logs","app.log");
+
+export const PATH_TO_LOGS_DIRECTORY = LOGS_DIRECTORY;
